More Graph and LineRenderer work
Added a line renderer class to be attached to the linerenderer object.

MyResources have been expanded with a colorpalette

GameController can now flip through playernames

Graph ALMOST supports multiple graphs. As of now, it supports two :(

This is more or less a backup before i go in and change some fundamentals in the graph class